move controller to legacy package
Moving haproxy ingress to controller-runtime in two steps. In the first
step, during v0.15, the controller will support two runtimes, and the
legacy version will be removed in v0.16. We're moving the legacy code
to a distinct package so the removal can be done later easily.
